[FEAT] make example work

+++ go_sdk/agoraservice/local_audio_track.go
@@ -26,7 +26,7 @@ func NewCustomPcmAudioTrack(pcmSender *AudioPcmDataSender) *LocalAudioTrack {
 }
 
 func NewCustomEncodedAudioTrack(encodedAudioSender *AudioEncodedFrameSender, mixMode int) *LocalAudioTrack {
-	cTrack := C.agora_service_create_custom_audio_track_encoded(agoraService.service, encodedAudioSender.cSender, mixMode)
+	cTrack := C.agora_service_create_custom_audio_track_encoded(agoraService.service, encodedAudioSender.cSender, C.int(mixMode))
 	if cTrack == nil {
 		return nil
 	}
